--- applications/accounts/templates/administrate/change_password.html
@@ -1,4 +1,0 @@
-{% extends 'layout.html' %}
-
-{% block body %}
-{% endblock %}
