--- applications/accounts/templates/administrate/change_permissions.html
@@ -1,21 +1,0 @@
-{% extends 'administrate/main_layout.html' %}
-
-{% block body %}
-    <div class="container">
-        <div class="panel panel-default">
-        <ul class="nav nav-tabs">
-            <li><a href="#">Главная</a></li>
-            <li class="active"><a href="#">Профиль</a></li>
-            <li><a href="#">Сообщение</a></li>
-        </ul>
-            <div class="panel-body">
-                <form action="">
-                    <p>sadasd</p>
-                    <p>sadasd</p>
-                    <p>sadasd</p>
-                    <p>sadasd</p>
-                </form>
-            </div>
-        </div>
-    </div>
-{% endblock %}
