--- applications/accounts/templates/administrate/main_layout.html
@@ -1,4 +1,0 @@
-{% extends 'layout.html' %}
-
-{% block body %}
-{% endblock %}
