+++ applications/infrastructure/admin.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals
 
+from django.conf import settings
 from django.contrib import admin
 
 from applications.infrastructure.models import Departament
@@ -18,3 +19,8 @@ class DepartamentAdmin(admin.ModelAdmin):
     prepopulated_fields = {'slug': ('title',)}
     inlines = [PositionInline, ]
     list_display = ('title', 'abbreviation', 'slug')
+
+
+default_site = 'Администрирование сайта'
+admin.site.site_header = getattr(settings, 'SITE_HEADER', default_site)
+admin.site.site_title = getattr(settings, 'SITE_TITLE', default_site)
