--- applications/infrastructure/urls.py
@@ -1,2 +1,0 @@
-from django.conf.urls import url
-
